requirement.txt
requests==2.32.5
BeautifulSoup4==4.13.2
selenium==4.41.0
webdriver-manager==4.0.2
langchain-core==1.0.0
langchain-openai==1.0.0
python-dotenv==1.1.0